chore: rename package from config to network
