+++ pkg/options/nats.go
@@ -3,16 +3,15 @@ package options
 import (
 	"fmt"
 	"log/slog"
-	"strings"
 
 	"github.com/nats-io/nats.go"
-	"github.com/synternet/data-layer-sdk/pkg/config"
+	"github.com/synternet/data-layer-sdk/pkg/network"
 )
 
 // MakeNats is a convenience function to create NATS connection from various parameters.
 func MakeNats(name, urls, userCreds, nkey, jwt, caCert, clientCert, clientKey string) (*nats.Conn, error) {
 	if urls == "" {
-		urls = strings.Join(config.DefaultNetworks.GetDefault().URLs, ",")
+		urls = network.Default().JoinURLs()
 		if urls == "" {
 			slog.Info("Will use NATS stub")
 			return nil, nil
